Change return type of elapsed_ms

+++ src/misc.h
@@ -102,7 +102,7 @@ extern HAMLIB_EXPORT(setting_t) rig_idx2setting(int i);
 
 extern HAMLIB_EXPORT(int) hl_usleep(useconds_t usec);
 
-extern HAMLIB_EXPORT(int) elapsed_ms(struct timespec *start, int start_flag);
+extern HAMLIB_EXPORT(double) elapsed_ms(struct timespec *start, int start_flag);
 
 #ifdef PRId64
 /** \brief printf(3) format to be used for long long (64bits) type */
